Added the Authorization attribute and JwtMiddleware
The current setup is the JwtMiddleware intercepts the incoming
requests and extracts the JwtToken and checks if token is valid
if yes then it adds User object to current context, if no then
nothing. the AuthorizeAttribute checks for every authorized method
if the current context object has User if not the request is
unauthorized.